newew

+++ People-Assignment/Controllers/PeopleController.cs
@@ -88,8 +88,17 @@ namespace People_Assignment.Controllers
         }
         public IActionResult Delete(int id)
         {
-            _peopleService.Remove(id);
-            return RedirectToAction(nameof(Person));
+            Person person = _peopleService.FindById(id);
+
+            if (person == null)
+            {
+                return RedirectToAction(nameof(Person));
+            }
+            else
+            {
+                _peopleService.Remove(id);
+            }
+            return View();
         }
 
         [HttpGet]
@@ -102,7 +111,8 @@ namespace People_Assignment.Controllers
         public IActionResult Search(string search)
         {
             List<Person> persons = _peopleService.Search(search);
-            if (persons != null)
+
+            if (search != null)
             {
                 return PartialView("_PersonRow", persons);
             }
